Rename path `distribution` -> `distributions`
