+++ pgmpy/distributions/categorical.py
